added validator to validate configurations
