+++ externals/adapters/postgres/index.js
@@ -3,8 +3,8 @@
 module.exports = (config) => {
     
     const pg = require("pg");
-    const validator = require('validate.js');
-    const dbAdapterError = require('./errors');
+    const validator = require('./validator');
+    const errors = require('./errors');
 
     // known error codes sent from postgres
     const ErrorCodes = {
@@ -13,16 +13,18 @@ module.exports = (config) => {
         CONNECTION_REFUSED: 'ECONNREFUSED'
     };
 
-    _checkConfig(config);
+    // check whether configuration values are valid
+    validator.validateConfig(config);
 
+    // create db pool
     let pool = _getPool(config);
 
     /**
      * Run a query.
      * 
-     * @param {*} query 
+     * @param {string} query 
      * @param {*} parameters 
-     * @param {*} resultCallback 
+     * @param {Function} resultCallback 
      */
     function query(query, parameters, resultCallback) {
 
@@ -66,72 +68,22 @@ module.exports = (config) => {
     }
 
     /**
-     * Check whether configurations are valid.
-     * 
-     * @param {*} config 
-     */
-    function _checkConfig(config) {
-
-        let rules = {
-            user: {
-                presence: true,
-                length: {
-                    minimum: 1
-                }
-            },
-            host: {
-                presence: true,
-                length: {
-                    minimum: 1
-                }
-            },
-            database: {
-                presence: true,
-                length: {
-                    minimum: 1
-                }
-            },
-            password: {
-                presence: true,
-                length: {
-                    minimum: 1  
-                }
-            },
-            port: {
-                presence: true,
-                numericality: {
-                    onlyInteger: true,
-                    greaterThan: 0
-                }
-            }
-        };
-
-        let details = validator.validate(config, rules);
-
-        if(details === undefined) {
-            return;
-        }
-
-        throw dbAdapterError.configError(details);
-    }
-
-    /**
      * Translate a postgres error to a standard DBAdapter error.
      * 
-     * @param {*} err 
+     * @param {Error} err 
      */
     function _getError(err) {
 
         switch(err.code) {
             
             case ErrorCodes.CONNECTION_ERROR:
-                return dbAdapterError.connectionError();
+                return errors.connectionError();
             case ErrorCodes.CONNECTION_REFUSED:
-                return dbAdapterError.connectionError();
+                return errors.connectionError();
             case ErrorCodes.DATABASE_NOT_FOUND:
-                return dbAdapterError.connectionError();
+                return errors.connectionError();
             default:
-                return dbAdapterError.unknownError();
+                return errors.unknownError();
         }
     }
